// types.ts
export type Role = 'Admin' | 'Manager' | 'Pracownik';

export interface User {
  id: number;
  name: string;
  title: string;
  email: string;
  avatar: string;
  role: Role;
  password: string;
}
// ...
export enum ProductionOrderStatus {
  ToDo = 'Do zrobienia',
  Done = 'Zrobione',
}

export type Palette = 'Atlas' | 'Sempre' | 'Teluria' | 'Tikkurila' | 'Inna';
export type ProductType = 'Tynk' | 'Farba';

export interface ProductionOrderItem {
  id: number;
  palette: Palette;
  productType: ProductType;
  base: string;
  color: string;
  capacity: string;
  quantity: number;
}

export interface ProductionOrder {
  id: number;
  orderNumber: string;
  creationDate: string;
  client: string;
  items: ProductionOrderItem[];
  createdBy: string;
  notes?: string;
  status: ProductionOrderStatus;
}